extract hash and hasher crates and re-export their contents in solana_program::hash
