+++ Cargo.toml
@@ -106,6 +106,8 @@ members = [
     "sdk/clock",
     "sdk/decode-error",
     "sdk/gen-headers",
+    "sdk/hash",
+    "sdk/hasher",
     "sdk/macro",
     "sdk/msg",
     "sdk/package-metadata-macro",
@@ -390,6 +392,8 @@ solana-genesis-utils = { path = "genesis-utils", version = "=2.1.0" }
 agave-geyser-plugin-interface = { path = "geyser-plugin-interface", version = "=2.1.0" }
 solana-geyser-plugin-manager = { path = "geyser-plugin-manager", version = "=2.1.0" }
 solana-gossip = { path = "gossip", version = "=2.1.0" }
+solana-hash = { path = "sdk/hash", version = "=2.1.0" }
+solana-hasher = { path = "sdk/hasher", version = "=2.1.0" }
 solana-inline-spl = { path = "inline-spl", version = "=2.1.0" }
 solana-lattice-hash = { path = "lattice-hash", version = "=2.1.0" }
 solana-ledger = { path = "ledger", version = "=2.1.0" }
