+++ sdk/hasher/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "solana-hasher"
+description = "Solana SHA256 hashing"
+documentation = "https://docs.rs/solana-hasher"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+sha2 = { workspace = true }
+solana-hash = { workspace = true }
+
+[target.'cfg(target_os = "solana")'.dependencies]
+solana-define-syscall = { workspace = true }
+
+[dev-dependencies]
+bs58 = { workspace = true }
