+++ sdk/program/src/syscalls/definitions.rs
@@ -1,7 +1,9 @@
 #[cfg(target_feature = "static-syscalls")]
 pub use solana_define_syscall::sys_hash;
-#[deprecated(since = "2.1.0", note = "Use `solana-msg::sol_log` instead.")]
+#[deprecated(since = "2.1.0", note = "Use `solana_msg::sol_log` instead.")]
 pub use solana_msg::sol_log;
+#[deprecated(since = "2.1.0", note = "Use solana_hasher::sol_sha256 instead")]
+pub use solana_hasher::sol_sha256;
 #[deprecated(
     since = "2.1.0",
     note = "Use `solana_program_memory::syscalls` instead"
@@ -24,7 +26,6 @@ define_syscall!(fn sol_log_compute_units_());
 define_syscall!(fn sol_log_pubkey(pubkey_addr: *const u8));
 define_syscall!(fn sol_create_program_address(seeds_addr: *const u8, seeds_len: u64, program_id_addr: *const u8, address_bytes_addr: *const u8) -> u64);
 define_syscall!(fn sol_try_find_program_address(seeds_addr: *const u8, seeds_len: u64, program_id_addr: *const u8, address_bytes_addr: *const u8, bump_seed_addr: *const u8) -> u64);
-define_syscall!(fn sol_sha256(vals: *const u8, val_len: u64, hash_result: *mut u8) -> u64);
 define_syscall!(fn sol_keccak256(vals: *const u8, val_len: u64, hash_result: *mut u8) -> u64);
 define_syscall!(fn sol_blake3(vals: *const u8, val_len: u64, hash_result: *mut u8) -> u64);
 define_syscall!(fn sol_invoke_signed_c(instruction_addr: *const u8, account_infos_addr: *const u8, account_infos_len: u64, signers_seeds_addr: *const u8, signers_seeds_len: u64) -> u64);
